// web/src/types/securevuConfig.ts
export interface SecureVuConfig {
// ...
  detectors: {
    coral: {
      device: string;
      model: {
        height: number;
        input_pixel_format: string;
        input_tensor: string;
        labelmap: Record<string, string>;
        labelmap_path: string | null;
        model_type: string;
        path: string;
        width: number;
      };
      type: string;
    };
  };
// ...
  model: {
    height: number;
    input_pixel_format: string;
    input_tensor: string;
    labelmap: Record<string, unknown>;
    labelmap_path: string | null;
    model_type: string;
    path: string | null;
    width: number;
    colormap: { [key: string]: [number, number, number] };
    attributes_map: { [key: string]: string[] };
    all_attributes: string[];
    plus?: {
      name: string;
      id: string;
      trainDate: string;
      baseModel: string;
      isBaseModel: boolean;
      supportedDetectors: string[];
      width: number;
      height: number;
    } | null;
  };
// ...
}